% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepthScaling.R
\name{depth.scaling}
\alias{depth.scaling}
\title{depth.scaling}
\usage{
depth.scaling(seuratob, assay = "RNA", counts.slot = "counts")
}
\arguments{
\item{seuratob}{Seurat object containing the raw transcript counts filtered for zero count genes. This object must contain all genes sequenced in the initial experiment.}

\item{assay}{Assay slot containing raw transcript counts (default "RNA").}

\item{counts.slot}{Slot within assay containing raw counts matrix (default "counts").}
}
\value{
Numeric vector containing the scaling coefficients for each cell.
}
\description{
depth.scaling
}
\examples{
depths <- depth.scaling(example.seurat)
}
