% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationMatrixFunctions.R
\name{FindCorrelationModules}
\alias{FindCorrelationModules}
\title{FindCorrelationModules}
\usage{
FindCorrelationModules(corr.matrix, num.steps = 4, weighted.clustering = FALSE)
}
\arguments{
\item{corr.matrix}{Sparse matrix containing the correlations of interest.}

\item{num.steps}{Integer. Number of steps used for walktrap clustering. (Default = 4)}

\item{weighted.clustering}{Boolean. Determines whether or not the walktrap will be performed on the correlations matrix or the adjacency matrix. (Default = FALSE)}
}
\value{
igraph communities object.
}
\description{
FindCorrelationModules
}
