% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationMatrixFunctions.R
\name{MergeSmallModules}
\alias{MergeSmallModules}
\title{MergeSmallModules}
\usage{
MergeSmallModules(corr.matrix, modules, min.size = 15)
}
\arguments{
\item{corr.matrix}{Sparse matrix containing the correlations of interest.}

\item{modules}{igraph communities object produced by FindCorrelationModules.}

\item{min.size}{Integer. Minimum size of output modules (if possible).}
}
\value{
igraph communities object containing the merged gene modules.
}
\description{
MergeSmallModules
}
