% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{StaticCorrelationPlot}
\alias{StaticCorrelationPlot}
\title{StaticCorrelationPlot}
\usage{
StaticCorrelationPlot(corr.matrix, highlight = NULL, modules = NULL)
}
\arguments{
\item{corr.matrix}{Sparse matrix containing the correlations of interest.}

\item{highlight}{Vector of gene symbols to be highlighted. If supplied, the relevant gene symbols will be written in red.}

\item{modules}{igraph communities object. If supplied, the nodes will change color and shape to indicate the module they belong to.}
}
\value{
ggraph plot of correlations.
}
\description{
StaticCorrelationPlot
}
