% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationMatrixFunctions.R
\name{SubsetCorrelationMatrix}
\alias{SubsetCorrelationMatrix}
\title{SubsetCorrelationMatrix}
\usage{
SubsetCorrelationMatrix(
  seuratObject,
  pCutoff = F,
  minPosCorr = F,
  minNegCorr = F,
  genes = F
)
}
\arguments{
\item{seuratObject}{Seurat object containing correlations matrix produced by BigSur.}

\item{pCutoff}{Boolean, Float. p-value threshold for correlations. Correlations with p-values larger than pCutoff will be set to 0.}

\item{minPosCorr}{Boolean, Float. Minimum positive correlation cutoff. Positive correlations with coefficients under minPosCorr will be set to 0.}

\item{minNegCorr}{Boolean, Float. Minimum negative correlation cutoff. Negative correlations with absolute values under abs(minNegCorr) will be set to 0.}

\item{genes}{Boolean, Character Vector. The set of genes for the correlation matrix to be subset to.}
}
\value{
Sparse matrix (dsCMatrix) containing correlations meeting subset criteria.
}
\description{
SubsetCorrelationMatrix
}
